가게 삭제 요청 추가

+++ src/main/java/com/depromeet/team5/service/impl/StoreServiceImpl.java
@@ -1,5 +1,6 @@
 package com.depromeet.team5.service.impl;
 
+import com.depromeet.team5.domain.DeleteRequestId;
 import com.depromeet.team5.domain.Image;
 import com.depromeet.team5.domain.Store;
 import com.depromeet.team5.domain.User;
@@ -7,7 +8,9 @@ import com.depromeet.team5.dto.StoreCardDto;
 import com.depromeet.team5.dto.StoreDto;
 import com.depromeet.team5.dto.UpdateDto;
 import com.depromeet.team5.exception.StoreNotFoundException;
+import com.depromeet.team5.exception.UserIdCheckException;
 import com.depromeet.team5.exception.UserNotFoundException;
+import com.depromeet.team5.repository.DeleteRepository;
 import com.depromeet.team5.repository.StoreRepository;
 import com.depromeet.team5.repository.UserRepository;
 import com.depromeet.team5.service.S3FileUploadService;
@@ -28,6 +31,7 @@ import java.util.stream.Collectors;
 public class StoreServiceImpl implements StoreService {
     private final UserRepository userRepository;
     private final StoreRepository storeRepository;
+    private final DeleteRepository deleteRepository;
     private final S3FileUploadService s3FileUploadService;
 
     @Override
@@ -35,9 +39,7 @@ public class StoreServiceImpl implements StoreService {
     public void saveStore(StoreDto storeDto, Long userId) {
         User user = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
         List<Image> image = convertImage(storeDto.getImage());
-
         Store store = Store.from(storeDto, image, user);
-        log.info(storeRepository.findAll().toString());
         storeRepository.save(store);
     }
 
@@ -74,9 +76,28 @@ public class StoreServiceImpl implements StoreService {
 
     @Override
     @Transactional
-    public void deleteStore(Long storeId) {
+    public void deleteStore(Long storeId, Long userId) {
+        userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
         Store store = storeRepository.findById(storeId).orElseThrow(StoreNotFoundException::new);
-        storeRepository.delete(store);
+        List<DeleteRequestId> deleteRequestId = store.getDeleteId();
+
+        if (deleteRepository.findByUserIdLike(userId).isPresent()) {
+            throw new UserIdCheckException();
+        }
+        DeleteRequestId deleteRequestId1 = new DeleteRequestId();
+        deleteRequestId1.setUserId(userId);
+
+        List<DeleteRequestId> deleteId = new ArrayList<>();
+        deleteId.addAll(deleteRequestId);
+        deleteId.add(deleteRequestId1);
+
+        store.getDeleteId().clear();
+        store.getDeleteId().addAll(deleteId);
+
+        if (deleteRequestId.size() == 5) {
+            storeRepository.delete(store);
+        }
+        
     }
